requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
colorama==0.4.6
distro==1.9.0
flet==0.25.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
oauthlib==3.2.2
openai==1.59.7
pillow==11.1.0
pydantic==2.10.5
pydantic_core==2.27.2
python-dotenv==1.0.1
repath==0.9.0
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
